pyproject.toml
[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "edupbl-backend"
version = "0.1.0"
description = "Sistema escolar modular - ocorrencias e frequencia"
requires-python = ">=3.11,<4.0"
# readme = "README.md"

dependencies = [
    "alembic (>=1.18.4,<2.0.0)",
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.31.0,<1.0.0)",
    "tzdata (>=2025.3,<2026.0)",
]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.15.1,<0.16.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "testcontainers (>=4.14.1,<5.0.0)",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = 'auto'
asyncio_default_fixture_loop_scope = "function"
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app/main.py'
test = 'pytest -vv -x -s --cov=app'
test_all = 'pytest -vv -rA --cov=app --cov-report=term-missing --cov-report=html'
test_resume = 'pytest -q -rA --tb=no'
post_test = 'coverage html'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
